perf: cs-etm: Set watermark for AUX trace
The default watermark is half of the total buffer size. In many cases,
the tool can not be notified with sufficient free space, especially when
profiling with small AUX buffer (e.g., 64KiB).

Setting watermark to quarter of the buffer to notifies the tool to read
data earlier and prevents the data loss.

Signed-off-by: Leo Yan <leo.yan@arm.com>