perf arm_spe: Consolidate operation types
Consolidate operation types in a more general way.

(a) The second-level types for memory and SIMD operations are classified
    by modules. E.g., an operation may relate to general register,
    SIMD/FP, SVE, etc.

(b) The associated information tells details. E.g., an operation is
    load or store, whether it is for data processing, etc.

Start the enum items for the second-level types from 8 to accommodate
more entries within a 32-bit integer.

As a first step, apply the macros to SVE operations (one for register
access and another for data processing).

Signed-off-by: Leo Yan <leo.yan@arm.com>