coresight: cti: Make spin lock usage consistent
The spin lock is currently acquired sometimes with IRQs disabled and
sometimes without. This leads to inconsistent semantics: the lock can be
either HARDIRQ-safe or HARDIRQ-unsafe, which may trigger lockdep
complaints.

Make the spin lock usage consistent by always without disabling IRQ.

Note: the CTI driver currently has a flaw that it does not use SMP call
for CPU-bound devices. Once the SMP call is supported in the CTI driver,
the locking scheme should consider to disable IRQs when taking the lock.

Signed-off-by: Leo Yan <leo.yan@arm.com>