coresight: syscfg: Use spinlock to protect active variables
To support dynamically enabling and disabling the Coresight path during
CPU idle, cscfg_config_sysfs_get_active_cfg() may be called by the ETM
driver in an atomic context. However, the function currently acquires a
mutex, which is a sleepable lock, and this is not allowed in atomic
context.

To make cscfg_config_sysfs_get_active_cfg() called in the idle flow,
replace the mutex with a raw spinlock to protect the active
variables.

Signed-off-by: Leo Yan <leo.yan@arm.com>