coresight: Save activated path into source device
Save activated path into the source device's coresight_device structure.
The path pointer will be used by later changes for controlling the path
during CPU idle.

The path pointer is assigned before setting the source device mode to
active, and it is cleared after the device is changed to an inactive
mode. So safe access to path pointers is guaranteed when the device is
in an active mode.

Signed-off-by: Leo Yan <leo.yan@arm.com>