coresight: cti: Refactor CPU PM notifier registration
CPU PM notifiers are a common mechanism and should not be registered
from probe, which is intended for device-specific initialization.

Move CPU PM notifier registration to module init so it is ready before
device probing, and switch to using cpuhp_setup_state_nocalls() to avoid
explicitly acquiring the CPU lock.  Unregister the notifiers in module
exit.

Signed-off-by: Leo Yan <leo.yan@arm.com>